requirements.txt
parallel-ssh
python-dotenv
pyyaml